requirements.txt
pandas
numpy
python-dotenv
faker
sqlalchemy 
psycopg2
ipykernel
matplotlib
seaborn
scikit-learn
shap
xgboost
catboost